+++ CHANGELOG.md
@@ -7,10 +7,12 @@
 - [TD-3929] Reindex the children of the structure domain when modifying
 
 ### Added
+
+- [TD-3453] Purge logically deleted structures
 - [TD-3906] Notes audit now includes status changes
 - [TD-3050] Show quality errors
 - [TD-3551] Restore rule implementations
-- [TD-3945] Created Grant entity
+- [TD-3945] Created `Grant` entity
 
 ## [4.24.0] 2021-07-13